requirements.txt
Django~=3.0.3
wheel~=0.34.2
mysqlclient~=1.4.6
djangorestframework~=3.11.0
djoser~=2.0.3
#markdown
#django-filter
